Move PubsubBoundedWriter outside of PubsubIO.
This should have no user-visible effects.
